public_projects/imgarden/requirements.txt
# 나비얌 챗봇 프로젝트 의존성

# 핵심 머신러닝 프레임워크
torch==2.3.0
transformers==4.40.0
accelerate==0.30.0
peft==0.11.0
bitsandbytes==0.43.0
datasets==2.19.0

# Vector DB 및 RAG 관련
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.24.3

# 텍스트 처리 및 NLP
sentencepiece==0.2.0
tokenizers==0.19.0

# 데이터 처리
pandas>=1.5.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
scikit-learn>=1.3.0

# 웹 및 API
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0

# 개발 및 테스트
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0

# 선택적 (GPU 환경용)
# torch-audio
# faiss-gpu>=1.7.4

# ChromaDB (선택적 Vector DB)
# chromadb>=0.4.0

# 상용 Vector DB (선택적)
# pinecone-client>=2.2.0
# weaviate-client>=3.20.0